(feat) moved studentHelpers into ModelHelpers
