+++ src/server/app.js
@@ -10,9 +10,9 @@ import Auth0Strategy from './auth/Auth0';
 import Auth0 from './auth/Auth0Helpers';
 import cloud from './cloudinary/cloudHelpers';
 import kairos from './kairosFR/kairosHelpers';
-import search from './db/search.js';
-import studentHelpers from './db/studentHelpers.js';
-import classHelpers from './db/ModelHelpers/classHelpers.js';
+import search from './db/search';
+import studentHelpers from './db/ModelHelpers/studentHelpers';
+import classHelpers from './db/ModelHelpers/classHelpers';
 import fileUpload from 'express-fileupload';
 import User from './db/userHelpers';
 import twilio from './twilio/twilioHelper';
